file-splitting & test run api in docker & code cleanup
