--- go/app/items.json
@@ -1,1 +1,0 @@
-{"items":[{"name":"","category":"","images":"","id":0},{"name":"jacket","category":"fashion","images":"","id":0},{"name":"jacket","category":"fashion","images":"","id":0},{"name":"jacket","category":"fashion","images":"","id":0},{"name":"smartphone","category":"","images":"","id":0},{"name":"smartphone","category":"","images":"","id":0},{"name":"smartphone","category":"","images":"","id":0},{"name":"smartphone","category":"","images":"","id":0},{"name":"smartphone","category":"","images":"","id":0},{"name":"a","category":"","images":"","id":0},{"name":"a","category":"","images":"","id":0},{"name":"a","category":"","images":"","id":0},{"name":"","category":"","images":"","id":0},{"name":"","category":"","images":"","id":0},{"name":"","category":"","images":"","id":0},{"name":"","category":"","images":"","id":0},{"name":"","category":"","images":"","id":0},{"name":"","category":"","images":"","id":0},{"name":"","category":"","images":"","id":0},{"name":"","category":"","images":"","id":0},{"name":"","category":"","images":"","id":0},{"name":"","category":"","images":"","id":0},{"name":"jacket","category":"fashion","images":"","id":0},{"name":"","category":"","images":"","id":0},{"name":"jacket","category":"","images":"","id":0},{"name":"jacket","category":"","images":"","id":0},{"name":"jacket","category":"","images":"","id":0},{"name":"","category":"","images":"","id":0},{"name":"jacket","category":"fashion","images":"","id":0},{"name":"jacket","category":"fashion","images":"","id":0},{"name":"jacket","category":"fashion","images":"","id":0},{"name":"jacket","category":"fashion","images":"","id":0},{"name":"jacket","category":"fashion","images":"","id":0},{"name":"jacket","category":"fashion","images":"","id":0},{"name":"jacket","category":"fashion","images":"","id":0},{"name":"jacket","category":"fashion","images":"","id":0},{"name":"jacket","category":"fashion","images":"ad55d25f2c10c56522147b214aeed7ad13319808d7ce999787ac8c239b24f71d","id":1},{"name":"jacket","category":"fashion","images":"ad55d25f2c10c56522147b214aeed7ad13319808d7ce999787ac8c239b24f71d","id":1}]}
